+++ CHANGELOG.md
@@ -58,6 +58,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 * Add Date Time Provider
 
+* Fakers for `Address`
+
 
 ### Updated